Update hmc_montecarlo.py
Streamlined code - it seems the issue is arising when calculating the jacobian. <A, x> is giving huge negative values which in turn is making the e^<A,x> equal 0 which in turn is making the entire estimate zero. But this raises the issue on why the value is so absurd in the first place and it might be an issue in the sampleing